feat: create LinkingScreen for email verification and remove legacy components
- Introduced LinkingScreen component to handle email verification via deep links, providing user feedback during the process.
- Removed outdated VerifyEmailScreen and related components to streamline the email verification flow.
- Updated deep link handling to route users to the new LinkingScreen for verification tasks.
